Local state opponent selector
Wiring up Apollo to use GraphQL for internal state management.
There's some aspects of the application state, that we don't need to
send to the GraphQL server. This will give us a client state for
tracking that, while still using the same GraphQL language.

Implementing it for the opponent selection. We'll store this in local
state and then send it along to the server later in the operation.