Dev tool for display local app state
The Apollo browser dev tools are pretty frustrating because they require
a manual reload to view the current state. This gives us a view into the
local app state alongside the run timers. Since this is a dev view of
the exhibit operations already, it makes sense to include this info in
the public view.