Fixing run date storage
There were a couple problems with the run dates.

The actual datetimes were getting generted on the component load, not
when the timers executed the mutation. Fixed this by moving the
variables within the mutation child call.

Also adding a date scalar type to our GraphQL server. GQL doesn't have a
native date type, but this module seems to be the standard way that
people are handling this. This allows us to pass the date around in the
same format that we're storing it within the DB. It's human readable, TZ
agnostic, and can be easily manipulated by Moment on the client. This
also gives us the milisecond precision we need for the run timers.