Pubish the run selected message
Modify the GraphQL client code to send a server request. Once this is
complete we update the local state using the Mutation component's update
prop.

On the server, our resolver doesn't preform any DB operations, just send
out the message that the race has been initiated with a new generated
date at the moment of initiation.