style: Organize documentation assets for beambreak
Moving some of the assets that Aaron added to break documentaion out
from any code. I'm adding a top level folder for `hardware`. Not sure if
this is the right place for this in the long-term, but it's code that's
part of this overal system that doesn't fit within the Lerna system.

I also renamed some files to make the case a bit more consistent across
some of our other files.

Issues: #186