feat: Change graphql server to apollo-server
The Apollo GraphQL server package has changed names and structure. We're
upgrading to the latest version, removing packages that are made
redundant by this upgrade, and modifying the server code to reflect the
new server functionality.

apollo-server-express covers the functionality of a couple of the
separate modules that we were previously using, making the server code
simpler. This includes a new GUI for the GraphQL server at
localhost:4000/graphql and a simpler way to handle publishing the pubsub
subscriptions for the messaging system.

In this commit, we also remove several modules that had been included in
the package.json for the server that either weren't being used or are
now unused with the new apollo-server-express module.

Issues: #208, #104