refactor: Eliminate Persona in favor or Person
Person is a simpler and more direct description of the data type being
entered. The record is a document of the "person" who is running. It
isn't an outward representation of their person, but rather their
actualy person data. Until we have the concept of a persona as well as
the data type of a person, adding persona just complicates the naming
unnescesarily.

Issues: #156