fix: Remove Avatar from GraphQL schema
This is an artifact of an older operation for the xMacroscope system.
In this context Avatar is synonymous with Opponent, or the character you
race against in the Run experience.

We decided to remove the opponent system from the xMacroscope for now.
This opponent racing is only done at SMM and won't be part of the
project until a later date.

Issues: #206, #156