feat: Remove dev messages and timer displays
In the exhibit run experience, the start kiosk doesn't show any timer
information. The timers are controlled on this system, but they are
displayed via the end line screen. Some timers like the timeout pause
and the delay for the audio display aren't displayed, because the audio
tells the visitor what is going on through the countdown.

Removed some of the language that explains the application state, that
isn't something we want to show visitors.

Modified the main state conditional so that you only show the runner's
icon when they are at the start line and likely to be looking at the
screen. We change it to a  message for the next runner, when the current
runner crosses the finish line.

Removed the touchscreen finish button. Finish triggers should only be
generated by the keypress generated by the end line sensor.

Issues: #206, #176, #213