refactor: Simplify the Run query
We had an old and out of date type called Run.

RunRecord is up to date with the correct data, but had a extraneous
name. I removed the old type and used the Run name in it's place.

Long story short, Run is the main data type for runs. RunInput is the
input type. Cleaned up the client queries and mutations. MAV team will
make their changes to relfect this.

Issues: #156