feat: Add person data to the run messages
We're removing the opponent select from the system, and adding person
data at the race initiation. This updates the subscriptions for the MAV
to expose that data.

The run-selected message is stil here, but isn't really used since
that's related to selecting an opponent.o

The run-initiated message now passes along the run data.

All messages now have a common format. There isn't a reason for them to
have a different data shape.

Issues: #156, #206, #171