fix: Modify build paths to reflect new directories
During development, the paths of several files have moved.

- Remove reference to base.css in the client-run app. This file is not
longer used and the reference is causing an error on build
- Update documentation to use .env vs .env.development. `.env` is
required when building the app for production so that the app knows the
GraphQL server location. The development distinction doesn't need to be
called out here. Users can customize during development on their own.
- Fix the server's express paths to reflect new app locations.