fix: Correct model case for graphql association
Sequelize will return included associations using the table names, which
are uppercase. Our GraphQL schema expects the associations to be
lowercase. So we need to set up an alias for Sequelize when returning
data with associations. I'd figure this out before but forgot to bring
it along in this project. This fixes #243 where the person data was
being lost in the find N runs query.

Issues: #243