feat: Enable subdirectory hosting for client-run
This sets up an basepath for the client app in production builds so that
it can properly be hosted from a subdirectory of the express server on
server-graphql.

Also cleaning up some of the server code in the graphql-server system,
to appropriately serve resources in development and production
environments.

Issues: #250, #251