feat: Cleanup untracked lock files
There is a desire by the CNS team to ignore the lock files in our
packages (#146).

Removing the previously tracked files from the git index and
updating the ignore file to explain this.

We use yarn still at SMM and so I'm also ignoring the yarn lock file
because it gets generated from time to time when we're jumping back and
forth between projects.