fix: Modify eslint rules to deal with Windows
Two small changes to address a windows conflict with the git deploy
process. In the future we can look at git config changes in the deploy
process to ensure that the line endings show up the same across all
envs. But for now, this allows the app to compile properly.