fix: Removing lock files again
There's been some discussion about lock file on issue #146.
I removed some lockfiles in a previous commit, but they seem to have
come back. I'm not able to track down where this happened. But they're
causing problems on our Windows install in the exhibit kiosks. Since we
have some modules that are optional on different OSes the lock files are
registering changes on the Windows machine.

I'd still rather we tracked the lock files, but until we can resolve
that we should be consistent in removing these from the repo.